Work on using floats in mouse picking, instead of ints

+++ data/shaders/node/node.frag
@@ -1,9 +1,9 @@
 #version 430 core
 
-in flat int v_entity_id;
+in flat float v_entity_id;
 
 layout(location = 0) out vec4 fragment_color;
-layout(location = 1) out int entity_id;
+layout(location = 1) out float entity_id;
 
 uniform vec4 u_color;
